Rename DashboadSection to AppInnerModule and move to /admin_ui/app project root, tidy up some comments
