+++ admin_ui/app/App.js
@@ -3,26 +3,15 @@ import { Router } from '@reach/router';
 import './main.css';
 import Store from './Store';
 
-/*
-  We want both the Login view and the Dashboard view to be part of the main bundle
-  (and thus appear immediatelly to the user once the bundle is loaded)
-  DashboardSection contains a nested <Router> which loads all
-  other pages using react-loadable
- */
 import LoginPage from './session/LoginPage';
-import DashboardSection from './dashboard/DashboardSection';
 import ForgotPasswordPage from './session/ForgotPasswordPage';
+import AppInnerModule from './AppInnerModule';
 
-
-
-/**
- * The top level component / wrapper
- */
 const App = () => {
   return (
     <Store.Provider>
       <Router>
-        <DashboardSection path="/*" />
+        <AppInnerModule path="/*" />
         <LoginPage path="/login" />
         <ForgotPasswordPage path="/forgot-password" />
       </Router>
